FIX: fix export import server list
